Update the base url to scrape and for the info urls
It seems like the council are now forcing people onto https, but it
doesn't cover the address '203.38.125.77', so when people click the
'read more information' link in PlanningAlerts it takes them to an
untrusted site page. See this example:

https://www.planningalerts.org.au/applications/534268

This updates the domain to one that works for both the scraping and the
info urls.